+++ src/core/statusbar/StatusBarManager.ts
